server/pyproject.toml
[project]
name = "idiotsms-account-server"
version = "1.0.0"
description = "Flask backend for IdiotsMS Account Management System"
authors = [
    {name = "IdiotsMS Team"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
    "flask-jwt-extended>=4.5.3",
    "pymysql>=1.1.0",
    "bcrypt>=4.0.1",
    "python-dotenv>=1.0.0",
    "flask-limiter>=3.5.0",
    "marshmallow>=3.20.1",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
